Chore: Add common library plumbing
Allow the client and server to share common code. In order to comply
with VSCode packaging, the code is compiled both in the server and the
client outputs. A package.json is only provided to install common
dependencies. Unit tests, linting and debugging configurations were
tested and updated to work with the new structure.